Split COM_ABI and WindowsRuntime_ABI targets/dirs again
