+++ Support/Sources/CMakeLists.txt
@@ -1,4 +1,6 @@
-add_subdirectory(ExportedABI)
+add_subdirectory(COM_ABI)
+add_subdirectory(WindowsRuntime_ABI)
 add_subdirectory(InternalABI)
+
 add_subdirectory(COM)
 add_subdirectory(WindowsRuntime)
