+++ Support/Sources/COM/CMakeLists.txt
@@ -1,3 +1,3 @@
 file(GLOB_RECURSE SOURCES *.swift)
 add_library(COM STATIC ${SOURCES})
-target_link_libraries(COM PRIVATE WindowsRuntime_ABI SWRT_InternalABI)
+target_link_libraries(COM PRIVATE COM_ABI SWRT_InternalABI)
