+++ Support/Sources/COM_ABI/CMakeLists.txt
@@ -1,0 +1,3 @@
+file(GLOB SOURCES *.c)
+add_library(COM_ABI STATIC ${SOURCES})
+target_include_directories(COM_ABI PUBLIC include)
