+++ Support/Sources/COM_ABI/combaseapi.c
