+++ Support/Sources/COM_ABI/include/SWRT/windows/BaseTsd.h
