+++ Support/Sources/COM_ABI/include/SWRT/windows/combaseapi.h
