+++ Support/Sources/COM_ABI/include/SWRT/windows/guiddef.h
