+++ Support/Sources/COM_ABI/include/SWRT/windows/objidl.h
