+++ Support/Sources/COM_ABI/include/SWRT/windows/oleauto.h
