+++ Support/Sources/InternalABI/CMakeLists.txt
@@ -1,4 +1,3 @@
-file(GLOB SOURCES *.c)
-add_library(SWRT_InternalABI STATIC ${SOURCES})
-target_include_directories(SWRT_InternalABI PUBLIC include)
-target_link_libraries(SWRT_InternalABI PUBLIC WindowsRuntime_ABI)
+add_library(SWRT_InternalABI INTERFACE)
+target_include_directories(SWRT_InternalABI INTERFACE include)
+target_link_libraries(SWRT_InternalABI INTERFACE COM_ABI WindowsRuntime_ABI)
