+++ Support/Sources/WindowsRuntime_ABI/CMakeLists.txt
@@ -1,0 +1,4 @@
+file(GLOB SOURCES *.c)
+add_library(WindowsRuntime_ABI STATIC ${SOURCES})
+target_include_directories(WindowsRuntime_ABI PUBLIC include)
+target_link_libraries(WindowsRuntime_ABI PUBLIC COM_ABI)
