+++ Support/Sources/WindowsRuntime_ABI/include/SWRT/windows/activation.h
