+++ Support/Sources/WindowsRuntime_ABI/include/SWRT/windows/inspectable.h
