+++ Support/Sources/WindowsRuntime_ABI/include/SWRT/windows/memorybuffer.h
