+++ Support/Sources/WindowsRuntime_ABI/include/SWRT/windows/restrictederrorinfo.h
