+++ Support/Sources/WindowsRuntime_ABI/include/SWRT/windows/roapi.h
@@ -9,6 +9,7 @@ typedef enum SWRT_RO_INIT_TYPE {
   SWRT_RO_INIT_MULTITHREADED = 1
 } SWRT_RO_INIT_TYPE;
 
+typedef SWRT_HResult(__stdcall *SWRT_DllGetActivationFactory)(SWRT_HString activatableClassId, SWRT_IActivationFactory** factory);
 SWRT_HResult SWRT_RoGetActivationFactory(SWRT_HString activatableClassId, SWRT_Guid* iid, void** factory);
 SWRT_HResult SWRT_RoInitialize(SWRT_RO_INIT_TYPE initType);
 void SWRT_RoUninitialize();
