+++ Support/Sources/WindowsRuntime_ABI/include/SWRT/windows/roerrorapi.h
