+++ Support/Sources/WindowsRuntime_ABI/include/SWRT/windows/weakreference.h
