+++ Support/Sources/WindowsRuntime_ABI/include/SWRT/windows/winstring.h
