+++ Support/Sources/WindowsRuntime_ABI/winstring.c
